Support build debug VM in CMake
